chore(shell-service)!: rename ServiceResult to ShellResult, move to shell module.
