+++ vsmtp-rule-engine/src/dsl/service/mod.rs
@@ -60,39 +60,3 @@ impl std::fmt::Display for Service {
         )
     }
 }
-
-/// Output generated by a service (shell)
-#[allow(clippy::module_name_repetitions)]
-#[derive(Debug, Clone, Copy)]
-pub struct ServiceResult {
-    // TODO: do we want ? ExitStatus or Output ? see Child::wait_with_output
-    status: std::process::ExitStatus,
-}
-
-impl ServiceResult {
-    pub const fn new(status: std::process::ExitStatus) -> Self {
-        Self { status }
-    }
-
-    pub fn has_code(self) -> bool {
-        self.get_code().is_some()
-    }
-
-    pub fn get_code(self) -> Option<i64> {
-        self.status.code().map(i64::from)
-    }
-
-    pub fn has_signal(self) -> bool {
-        self.get_signal().is_some()
-    }
-
-    pub fn get_signal(self) -> Option<i64> {
-        std::os::unix::prelude::ExitStatusExt::signal(&self.status).map(i64::from)
-    }
-}
-
-impl std::fmt::Display for ServiceResult {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_fmt(format_args!("{}", self.status))
-    }
-}
