+++ vsmtp-rule-engine/src/dsl/service/shell.rs
@@ -15,7 +15,6 @@
  *
 */
 
-use crate::dsl::service::ServiceResult;
 use crate::log_channels;
 use crate::{dsl::service::Service, modules::EngineResult};
 use rhai::EvalAltResult;
@@ -68,7 +67,7 @@ pub fn parse_shell_service(
 /// * if the group used to launch commands is not found.
 /// * if the shell service failed to spawn.
 /// * if the shell returned an error.
-pub fn run(service: &Service) -> anyhow::Result<ServiceResult> {
+pub fn run(service: &Service) -> anyhow::Result<ShellResult> {
     if let Service::UnixShell {
         timeout,
         command,
@@ -114,8 +113,44 @@ pub fn run(service: &Service) -> anyhow::Result<ServiceResult> {
             Err(err) => anyhow::bail!("shell unexpected error: {err:?}"),
         };
 
-        Ok(ServiceResult::new(status))
+        Ok(ShellResult::new(status))
     } else {
         anyhow::bail!("only a shell service can use the 'run' function")
+    }
+}
+
+/// Output generated by a service (shell)
+#[allow(clippy::module_name_repetitions)]
+#[derive(Debug, Clone, Copy)]
+pub struct ShellResult {
+    // TODO: do we want ? ExitStatus or Output ? see Child::wait_with_output
+    status: std::process::ExitStatus,
+}
+
+impl ShellResult {
+    pub const fn new(status: std::process::ExitStatus) -> Self {
+        Self { status }
+    }
+
+    pub fn has_code(self) -> bool {
+        self.get_code().is_some()
+    }
+
+    pub fn get_code(self) -> Option<i64> {
+        self.status.code().map(i64::from)
+    }
+
+    pub fn has_signal(self) -> bool {
+        self.get_signal().is_some()
+    }
+
+    pub fn get_signal(self) -> Option<i64> {
+        std::os::unix::prelude::ExitStatusExt::signal(&self.status).map(i64::from)
+    }
+}
+
+impl std::fmt::Display for ShellResult {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.write_fmt(format_args!("{}", self.status))
     }
 }
